[dhctl] Refactor resources checking
