+++ dhctl/pkg/apis/v1/nodegroup.go
@@ -1,4 +1,4 @@
-package resources
+package v1
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
